+++ src/main/java/com/agilityhub/core/configuration/MongoConfiguration.java
@@ -1,0 +1,15 @@
+package com.agilityhub.core.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.MongoDatabaseFactory;
+import org.springframework.data.mongodb.MongoTransactionManager;
+
+@Configuration(proxyBeanMethods = false)
+public class MongoConfiguration {
+
+    @Bean
+    MongoTransactionManager transactionManager(MongoDatabaseFactory databaseFactory) {
+        return new MongoTransactionManager(databaseFactory);
+    }
+}